+++ src/modules/user/useCases/mappers/__tests__/mapUserEntityToResponse.test.ts
