+++ src/modules/user/useCases/mappers/mapUserEntityToResponse.ts
@@ -1,5 +1,5 @@
 import { User } from '../../entities/User';
-import { UserResponse } from '../../useCases/interfaces/UserResponse';
+import { UserResponse } from '../interfaces/UserResponse';
 
 export const mapUserEntityToResponse = (user: User): UserResponse => {
     const { name, email, cpf, hasAddress, password, id } = user.getUser();
